use configured min_pressure

+++ github-runner-manager/tests/unit/manager/test_pressure_reconciler.py
@@ -44,7 +44,7 @@ class _FakePlanner:
 
     def __init__(
         self,
-        stream_updates: list[int] | None = None,
+        stream_updates: list[float] | None = None,
         stream_raises: bool = False,
     ):
         """Initialize with configurable stream behavior."""
